try to fix MSVC compile warning

+++ arangod/Agency/Agent.cpp
@@ -1258,7 +1258,7 @@ write_ret_t Agent::write(query_t const& query, WriteMode const& wmode) {
       indices.insert(indices.end(), tmp.begin(), tmp.end());
     }
     _write_hist_msec.count(
-      duration<double,std::milli>(high_resolution_clock::now()-start).count());
+      duration<float, std::milli>(high_resolution_clock::now()-start).count());
   }
 
   // Maximum log index
